[Refactoring] Move ExtractFunction to its own file
