+++ lib/Refactoring/CMakeLists.txt
@@ -6,6 +6,7 @@ add_swift_host_library(swiftRefactoring STATIC
   ConvertToSwitchStmt.cpp
   ConvertToTernaryExpr.cpp
   ExtractRepeatedExpr.cpp
+  ExtractFunction.cpp
   MoveMembersToExtension.cpp
   Refactoring.cpp
   Renamer.cpp
