+++ lib/Refactoring/Utils.cpp
@@ -11,6 +11,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "Utils.h"
+#include "swift/AST/NameLookup.h"
 
 using namespace swift::refactoring;
 
@@ -43,4 +44,15 @@ swift::refactoring::correctNameInternal(ASTContext &Ctx, StringRef Name,
       break;
   }
   return Ctx.getIdentifier((llvm::Twine(Name) + SuffixToUse).str()).str();
+}
+
+llvm::StringRef swift::refactoring::correctNewDeclName(DeclContext *DC,
+                                                       StringRef Name) {
+
+  // Collect all visible decls in the decl context.
+  llvm::SmallVector<ValueDecl *, 16> AllVisibles;
+  VectorDeclConsumer Consumer(AllVisibles);
+  ASTContext &Ctx = DC->getASTContext();
+  lookupVisibleDecls(Consumer, DC, true);
+  return correctNameInternal(Ctx, Name, AllVisibles);
 }
